services/telegram-webhook/requirements.txt
Flask
requests
cloudinary
firebase-admin
python-dotenv
google-genai
APScheduler
